change offset to be an unsigned, as it's always <= GIMLI_RATE
fix conversion warnings in gimli_absorb_byte and gimli_squeeze_byte

+++ include/lithium/gimli_state.h
@@ -16,7 +16,7 @@
 typedef struct
 {
     uint32_t state[GIMLI_WORDS];
-    size_t offset;
+    unsigned offset;
 } gimli_state;
 
 /* cffi:end */
